old swarm/swarm module/Red_team/requirements.txt
# Core
pydantic>=2.0,<3.0
pydantic-settings>=2.0,<3.0

# AI / LLM
ollama>=0.4,<1.0
openai>=1.0,<2.0
langgraph==0.2.48
langchain==0.3.19
langchain-core>=0.3.35,<1.0
langchain-openai==0.2.14

# Infrastructure
redis>=5.0,<6.0

# Tools & Sandbox
docker>=7.0,<8.0
httpx>=0.28,<1.0
pyyaml>=6.0,<7.0
gitpython>=3.1,<4.0

# API Server
fastapi>=0.115,<1.0
uvicorn>=0.32,<1.0

# Testing
pytest>=8.0,<9.0
pytest-asyncio>=0.24,<1.0

# Optional - Episodic Memory (Qdrant)
qdrant-client>=1.12.0,<2.0
